// src/interfaces/pokemonInterfaces.ts
export interface SimplePokemon {
    id:string;
    name: string;
    picture: string;
    color?:string
}